+++ uat/artifacts/src/main/java/com/aws/greengrass/Main.java
