+++ uat/artifacts/src/main/java/com/aws/greengrass/artifacts/logGenerator.java
