+++ uat/artifacts/src/test/java/com/aws/greengrass/LogGeneratorTest.java
