+++ uat/features/pom.xml
@@ -9,8 +9,8 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.aws.greengrass</groupId>
-    <artifactId>log-manager-testing-features</artifactId>
+    <artifactId>log-manager-features</artifactId>
+    <groupId>com.aws.greengrass.testing</groupId>
     <version>1.0-SNAPSHOT</version>
     <properties>
         <maven.compiler.source>1.8</maven.compiler.source>
@@ -79,27 +79,6 @@
     <build>
         <plugins>
             <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-antrun-plugin</artifactId>
-                <version>3.0.0</version>
-                <executions>
-                    <execution>
-                        <id>copy-artifact-to-classpath</id>
-                        <phase>package</phase>
-                        <goals>
-                            <goal>run</goal>
-                        </goals>
-                        <configuration>
-                            <skip>${skipTests}</skip>
-                            <target name="copy only uat jar">
-                                <copy file="${project.basedir}/target/${project.artifactId}-${project.version}.jar"
-                                      tofile="${project.basedir}/target/classes/greengrass/artifacts/componentArtifact.jar"/>
-                            </target>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
                 <artifactId>maven-resources-plugin</artifactId>
                 <version>2.6</version>
                 <executions>
@@ -113,7 +92,7 @@
                             <outputDirectory>${basedir}/target/classes/local-store/recipes</outputDirectory>
                             <resources>
                                 <resource>
-                                    <directory>${basedir}/../../uat/custom-components/target/classes/recipes</directory>
+                                    <directory>${basedir}/../../uat/features/target/classes/greengrass/recipes</directory>
                                 </resource>
                             </resources>
                         </configuration>
@@ -128,9 +107,27 @@
                             <outputDirectory>${basedir}/target/classes/local-store/artifacts/</outputDirectory>
                             <resources>
                                 <resource>
-                                    <directory>${basedir}/../../uat/custom-components/target/</directory>
+                                    <directory>${basedir}/../../uat/artifacts/target/</directory>
                                     <includes>
-                                        <include>custom-components.jar</include>
+                                        <include>LocalArtifacts.jar</include>
+                                    </includes>
+                                </resource>
+                            </resources>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>copy-component-artifact</id>
+                        <phase>process-classes</phase>
+                        <goals>
+                            <goal>copy-resources</goal>
+                        </goals>
+                        <configuration>
+                            <outputDirectory>${project.basedir}/target/classes/greengrass/artifacts/</outputDirectory>
+                            <resources>
+                                <resource>
+                                    <directory>../../target/</directory>
+                                    <includes>
+                                        <include>LogManager.jar</include>
                                     </includes>
                                 </resource>
                             </resources>
@@ -149,16 +146,21 @@
                             <goal>shade</goal>
                         </goals>
                         <configuration>
-                            <finalName>greengrass-log-manager-testing-features</finalName>
+                            <finalName>log-manager-uats</finalName>
                             <transformers>
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.ApacheLicenseResourceTransformer"/>
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                                <transformer
+                                        implementation="org.apache.maven.plugins.shade.resource.ApacheLicenseResourceTransformer"/>
+                                <transformer
+                                        implementation="org.apache.maven.plugins.shade.resource.ServicesResourceTransformer"/>
+                                <transformer
+                                        implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                                     <mainClass>com.aws.greengrass.testing.launcher.TestLauncher</mainClass>
                                 </transformer>
-                                <transformer implementation="org.apache.maven.plugins.shade.resource.IncludeResourceTransformer">
+                                <transformer
+                                        implementation="org.apache.maven.plugins.shade.resource.IncludeResourceTransformer">
                                     <resource>greengrass/artifacts/componentArtifact.jar</resource>
-                                    <file>${project.basedir}/target/classes/greengrass/artifacts/componentArtifact.jar</file>
+                                    <file>${project.basedir}/target/classes/greengrass/artifacts/componentArtifact.jar
+                                    </file>
                                 </transformer>
                             </transformers>
                         </configuration>
