+++ uat/features/src/main/java/com/aws/greengrass/testing/resources/cloudwatch/CloudWatchLogStreamModel.java
@@ -3,12 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.resources;
+package com.aws.greengrass.testing.resources.cloudwatch;
 
 import com.aws.greengrass.logging.api.Logger;
 import com.aws.greengrass.logging.impl.LogManager;
 import com.aws.greengrass.testing.api.model.TestingModel;
 import com.aws.greengrass.testing.resources.AWSResource;
+import com.aws.greengrass.testing.resources.cloudwatch.CloudWatchLogStream;
 import org.immutables.value.Value;
 import software.amazon.awssdk.services.cloudwatchlogs.CloudWatchLogsClient;
 import software.amazon.awssdk.services.cloudwatchlogs.model.DeleteLogStreamRequest;
