+++ uat/features/src/main/java/com/aws/greengrass/testing/resources/cloudwatch/CloudWatchLogsLifecycle.java
@@ -3,10 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.resources;
+package com.aws.greengrass.testing.resources.cloudwatch;
 
 import com.aws.greengrass.testing.resources.AWSResourceLifecycle;
 import com.aws.greengrass.testing.resources.AbstractAWSResourceLifecycle;
+import com.aws.greengrass.testing.resources.cloudwatch.CloudWatchLogStreamSpec;
 import com.google.auto.service.AutoService;
 import software.amazon.awssdk.services.cloudwatchlogs.CloudWatchLogsClient;
 import software.amazon.awssdk.services.cloudwatchlogs.model.DescribeLogStreamsRequest;
