+++ uat/features/src/main/resources/greengrass/features/log-manager-1.feature
