+++ uat/pom.xml
@@ -15,10 +15,10 @@
     <packaging>pom</packaging>
     <modules>
         <module>
-            custom-components
+           artifacts
         </module>
         <module>
-            testing-features
+          features
         </module>
     </modules>
     <properties>
